requirements.txt
Package                       Version
----------------------------- ---------
asgiref                       3.3.1
certifi                       2020.12.5
chardet                       3.0.4
Django                        3.1.4
django-environ                0.4.5
djangorestframework           3.12.2
djangorestframework-simplejwt 4.6.0
idna                          2.10
pip                           20.3.1
PyJWT                         1.7.1
pytz                          2020.4
requests                      2.25.0
setuptools                    41.2.0
sqlparse                      0.4.1
urllib3                       1.26.2